requirements.txt
Flask>=2.0.0
azure-storage-blob>=12.0.0
gunicorn
